+++ packages/react-tinacms-inline/src/styles/index.ts
@@ -15,5 +15,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 */
-export * from './input-focus-wrapper'
 export * from './focus-ring'