+++ packages/next-tinacms-json/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.27.1](https://github.com/tinacms/tinacms/compare/v0.27.0...v0.27.1) (2020-08-10)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+# [0.27.0](https://github.com/tinacms/tinacms/compare/v0.26.0...v0.27.0) (2020-08-10)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+# [0.26.0](https://github.com/tinacms/tinacms/compare/v0.25.0...v0.26.0) (2020-08-03)
+
+
+### Bug Fixes
+
+* **next-tinacms-json:** useJsonForm#loadInitialValues does not run when cms is disabled ([9fbd8e8](https://github.com/tinacms/tinacms/commit/9fbd8e83d1765a97b747fca441869538137488bb))
+
+
+
+
+
 ## [0.3.24-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.23...next-tinacms-json@0.3.24-alpha.0) (2020-07-15)
 
 **Note:** Version bump only for package next-tinacms-json